// src/models/menu/MenuItem.enum.ts
export enum MenuItemEnum {
  DASHBOARD = 'DASHBOARD',
  DEV_PATTERNS = 'DEV_PATTERNS',
  PATHS = 'PATHS',
  PROCESSES = 'PROCESSES',
  QUESTION_BANK = 'QUESTION_BANK',
  SETTINGS = 'SETTINGS',
  SURVEYS = 'SURVEYS',
  TESTS = 'TESTS',
}

export enum MenuSectionEnum {
  DEVELOPER = 'DEVELOPER',
  MODULES = 'MODULES',
  OVERVIEW = 'OVERVIEW',
  SYSTEM = 'SYSTEM',
}